moved process_info into process
